+++ file-sharing/src/tabs/iSCSI/ui/screens/virtualDevice/VirtualDeviceTable.vue
@@ -1,7 +1,16 @@
 <template>
   <CardContainer>
     <template v-slot:header> Devices </template>
-    <div class="overflow-hidden" :style="{
+    <div v-if="useUserSettings().value.iscsi.clusteredServer" class="overflow-hidden" :style="{
+      'max-height': showAddDevice ? '1500px' : '0',
+      transition: showAddDevice ? 'max-height 0.5s ease-in' : 'max-height 0.5s ease-out',
+    }">
+      <div class="card">
+        <RBDDeviceCreationScreen @created="selectVirtualDevice" @close="showAddDevice = false" />
+      </div>
+
+    </div>
+    <div v-else class="overflow-hidden" :style="{
       'max-height': showAddDevice ? '1500px' : '0',
       transition: showAddDevice ? 'max-height 0.5s ease-in' : 'max-height 0.5s ease-out',
     }">
@@ -23,19 +32,31 @@
               <th scope="col">File Path</th>
               <th scope="col">Block Size</th>
               <th scope="col">Type</th>
-              <th v-if="useUserSettings().value.iscsi.clusteredServer" scope=" col">In Use</th>
+              <th scope=" col">In Use</th>
+              <th v-if="useUserSettings().value.iscsi.clusteredServer" scope="col"> Node IP
+              </th>
               <th scope="col" class="flex flex-row justify-end">
                 <span class="sr-only">Delete</span>
                 <button @click="showAddDevice = !showAddDevice">
                   <PlusIcon class="size-icon icon-default" />
                 </button>
               </th>
+
             </tr>
           </template>
           <template #tbody>
-            <VirtualDeviceEntry v-for="(device, index) in virtualDevices" :key="index" :device="device"
-              @deleteDevice="actions.refreshDevices" />
+            <template v-if="!fetchingDevices">
+              <template v-for="(device, index) in virtualDevices" :key="index">
+                <VirtualDeviceEntry :device="device" @deleteDevice="actions.refreshDevices"
+                  @update="actions.refreshDevices" />
+              </template>
+            </template>
+            <tr v-else>
+              <td colspan="6" class="p-4 text-gray-500 text-sm text-center">Loading devices...</td>
+            </tr>
           </template>
+
+
         </Table>
       </div>
     </div>
@@ -44,7 +65,7 @@
 
 <script setup lang="ts">
 import { CardContainer, wrapActions, Table } from "@45drives/houston-common-ui";
-import { inject, ref, watch, type Ref } from "vue";
+import { inject, ref, watch, type Ref, computed } from "vue";
 import { VirtualDevice } from "@/tabs/iSCSI/types/VirtualDevice";
 import { PlusIcon } from "@heroicons/vue/24/solid";
 import VirtualDeviceEntry from "../virtualDevice/VirtualDeviceEntry.vue";
@@ -53,14 +74,17 @@ import { useUserSettings } from "@/common/user-settings";
 import type { ResultAsync } from "neverthrow";
 import type { ISCSIDriver } from "@/tabs/iSCSI/types/drivers/ISCSIDriver";
 import type { ProcessError } from "@45drives/houston-common-lib";
+import RBDDeviceCreationScreen from "../radosBlockDeviceManagement/RBDDeviceCreationScreen.vue";
 
 const showAddDevice = ref(false);
 
 const driver = inject<ResultAsync<ISCSIDriver, ProcessError>>("iSCSIDriver")!;
 
 const virtualDevices = inject<Ref<VirtualDevice[]>>("virtualDevices")!;
+const showAddRBD = ref(false);
 
 const forceRefreshRecords = inject<Record<string, boolean>>("forceRefreshRecords")!;
+const fetchingDevices = ref<boolean>(false);
 
 watch(forceRefreshRecords, () => {
   if (forceRefreshRecords["devices"]) {
@@ -69,12 +93,36 @@ watch(forceRefreshRecords, () => {
   }
 });
 
+
+const emit = defineEmits<{
+  (e: 'update'): void;
+  (e: 'selectDevice', value: VirtualDevice): void;
+}>();
+
+
+const selectVirtualDevice = (device: VirtualDevice) => {
+  if (virtualDevices.value.find((existingDevice) => (existingDevice.deviceName === device.deviceName)) === undefined) {
+    return driver.andThen((driver) => driver.addVirtualDevice(device))
+      .andThen(() => actions.refreshDevices())           // refresh list
+      .map(() => {
+        emit('selectDevice', device)
+        showAddDevice.value = false;
+      });
+  }
+  return undefined;
+}
 const refreshDevices = () => {
-  return driver.andThen((driver) => {
-    return driver.getVirtualDevices().map((devices) => {
-      virtualDevices.value = devices;
-    });
-  });
+  fetchingDevices.value = true;
+  return driver
+    .andThen((driver) =>
+      driver.getVirtualDevices().map((devices) => {
+        virtualDevices.value = devices;
+        fetchingDevices.value = false;
+
+      })
+    )
+
+
 };
 
 const actions = wrapActions({ refreshDevices: refreshDevices });