+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## Cockpit File Sharing 4.3.2-1
+## Cockpit File Sharing 4.4.1-10
 
-* Fix bug causing nfs server crash during export removal with ceph by removing ceph remount after removing share instead of before
+* building package internal for repo